- migrated HF usage to HuggingFaceEndpoint as HuggingFaceHub was deprecated - moved collect_urls func to utils.py
